# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Books API",
    "version": "1.0.0",
    "description": "A simple API for working with books"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "summary": "Get all books",
        "tags": [
          "Books"
        ],
        "responses": {
          "200": {
            "description": "A list of books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "6a972bfc6201314a5ceb8ec5"
                      },
                      "id": {
                        "type": "string",
                        "example": "b1"
                      },
                      "author": {
                        "type": "string",
                        "example": "George Orwell"
                      },
                      "title": {
                        "type": "string",
                        "example": "1984"
                      },
                      "publicationDate": {
                        "type": "string",
                        "example": "1949-06-08"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "summary": "Get a single book by id",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The book's custom id (e.g. \"b1\")"
          }
        ],
        "responses": {
          "200": {
            "description": "The matching book",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "6a972bfc6201314a5ceb8ec5"
                    },
                    "id": {
                      "type": "string",
                      "example": "b1"
                    },
                    "author": {
                      "type": "string",
                      "example": "George Orwell"
                    },
                    "title": {
                      "type": "string",
                      "example": "1984"
                    },
                    "publicationDate": {
                      "type": "string",
                      "example": "1949-06-08"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No book found with the given id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Book not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/authors": {
      "get": {
        "summary": "Get all authors",
        "tags": [
          "Authors"
        ],
        "responses": {
          "200": {
            "description": "A list of authors"
          },
          "500": {
            "description": "Unable to retrieve authors"
          }
        }
      },
      "post": {
        "summary": "Create a new author",
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "name",
                  "birthYear"
                ],
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "birthYear": {
                    "type": "integer"
                  }
                }
              },
              "example": {
                "id": "a4",
                "name": "Example Author",
                "birthYear": 1980
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created author"
          },
          "400": {
            "description": "A required field is missing, or the id already exists"
          },
          "500": {
            "description": "Unable to create author"
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "summary": "Get an author by id",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The matching author"
          },
          "404": {
            "description": "Author not found"
          },
          "500": {
            "description": "Unable to retrieve author"
          }
        }
      },
      "put": {
        "summary": "Update an existing author",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "birthYear"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "birthYear": {
                    "type": "integer"
                  }
                }
              },
              "example": {
                "name": "Updated Author",
                "birthYear": 1981
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated author"
          },
          "400": {
            "description": "A required field is missing"
          },
          "404": {
            "description": "Author not found"
          },
          "500": {
            "description": "Unable to update author"
          }
        }
      },
      "delete": {
        "summary": "Delete an existing author",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Author deleted successfully"
          },
          "404": {
            "description": "Author not found"
          },
          "409": {
            "description": "Cannot delete an author who still has books"
          },
          "500": {
            "description": "Unable to delete author"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}